+++ packages/models/src/Domain/Syncable/Note/index.ts
@@ -1,4 +1,3 @@
 export * from './Note'
 export * from './NoteMutator'
 export * from './NoteContent'
-export * from './NoteBlocks'