+++ packages/web/src/javascripts/Components/BlockEditor/BlockEditor.tsx
@@ -2,36 +2,37 @@ import { WebApplication } from '@/Application/Application'
 import { SNNote } from '@standardnotes/snjs'
 import { FunctionComponent, useCallback, useRef } from 'react'
 import { BlockEditorController } from './BlockEditorController'
-import { AddBlockButton } from './BlockMenu/AddButton'
-import { MultiBlockRenderer } from './BlockRender/MultiBlockRenderer'
-import { BlockOption } from './BlockMenu/BlockOption'
+import { BlocksEditor } from '@standardnotes/blocks-editor'
+
+const StringEllipses = '...'
+const NotePreviewCharLimit = 160
 
 type Props = {
   application: WebApplication
   note: SNNote
 }
+
 export const BlockEditor: FunctionComponent<Props> = ({ note, application }) => {
   const controller = useRef(new BlockEditorController(note, application))
 
-  const onSelectOption = useCallback(
-    (option: BlockOption) => {
-      void controller.current.addNewBlock(option)
+  const handleChange = useCallback(
+    (value: string) => {
+      const content = value
+      const truncate = content.length > NotePreviewCharLimit
+      const substring = content.substring(0, NotePreviewCharLimit)
+      const previewPlain = substring + (truncate ? StringEllipses : '')
+      void controller.current.save({ text: content, previewPlain: previewPlain, previewHtml: undefined })
     },
     [controller],
   )
 
   return (
-    <div className="w-full">
-      <AddBlockButton application={application} onSelectOption={onSelectOption} />
-      {note.blocksItem && (
-        <MultiBlockRenderer
-          controller={controller.current}
-          key={note.uuid}
-          blocksItem={note.blocksItem}
-          note={note}
-          application={application}
-        />
-      )}
+    <div className="relative h-full w-full p-5">
+      <BlocksEditor
+        onChange={handleChange}
+        initialValue={note.content.text}
+        className="relative relative resize-none text-base focus:shadow-none focus:outline-none"
+      />
     </div>
   )
 }