+++ packages/web/src/javascripts/Components/BlockEditor/BlockRender/MaxBlockHeight.tsx
@@ -1,3 +1,0 @@
-export const MaxBlockHeight: Record<string, number> = {
-  'org.standardnotes.standard-sheets': 300,
-}