+++ packages/encryption/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.19.7](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.19.6...@standardnotes/encryption@1.19.7) (2022-11-07)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
 ## [1.19.6](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.19.5...@standardnotes/encryption@1.19.6) (2022-11-06)
 
 **Note:** Version bump only for package @standardnotes/encryption