+++ packages/filepicker/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.25.7](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.25.6...@standardnotes/filepicker@1.25.7) (2022-11-07)
+
+**Note:** Version bump only for package @standardnotes/filepicker
+
 ## [1.25.6](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.25.5...@standardnotes/filepicker@1.25.6) (2022-11-06)
 
 **Note:** Version bump only for package @standardnotes/filepicker