+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.213](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.212...@standardnotes/releases@1.3.213) (2022-11-07)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.3.212](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.211...@standardnotes/releases@1.3.212) (2022-11-07)
 
 **Note:** Version bump only for package @standardnotes/releases