+++ packages/ui-services/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.6](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.13.5...@standardnotes/ui-services@1.13.6) (2022-11-07)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
 ## [1.13.5](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.13.4...@standardnotes/ui-services@1.13.5) (2022-11-06)
 
 **Note:** Version bump only for package @standardnotes/ui-services