+++ packages/web/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.93.13](https://github.com/standardnotes/app/compare/@standardnotes/web@3.93.12...@standardnotes/web@3.93.13) (2022-11-07)
+
+**Note:** Version bump only for package @standardnotes/web
+
 ## [3.93.12](https://github.com/standardnotes/app/compare/@standardnotes/web@3.93.11...@standardnotes/web@3.93.12) (2022-11-07)
 
 ### Bug Fixes